requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-cors-headers==3.14.0
django-countries==7.5.1
django-phone-field==1.8.1
django-simple-history==3.3.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.5.0
